webapi/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
billiard==3.6.4.0
boto3==1.40.65
botocore==1.40.65
cachetools==5.5.2
celery==5.2.7
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
deepeval==3.7.0
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
et_xmlfile==2.0.0
execnet==2.1.1
fastapi==0.121.0
fastuuid==0.14.0
filelock==3.20.0
frozenlist==1.6.2
fsspec==2024.12.0
google-auth==2.40.2
google-genai==1.18.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.72.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huey==2.5.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
kuzu==0.11.3
langchain==1.0.3
langchain-aws==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==0.3.79
langchain-memgraph==0.1.9
langchain-neo4j==0.6.0
langchain-openai==0.3.19
langchain-qdrant==0.2.0
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.40
litellm==1.79.1
loadenv==0.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.20.0
mdurl==0.1.2
mem0ai==1.0.0
memgraph-toolbox==0.1.6
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
neo4j==5.27.0
neo4j-graphrag==1.10.1
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
ollama==0.6.0
openai==2.7.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.0
pillow==12.0.0
pluggy==1.6.0
portalocker==2.10.1
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyfiglet==1.0.4
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.1.3
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-repeat==0.9.4
pytest-rerunfailures==12.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.3
qdrant-client==1.14.2
rank-bm25==0.2.2
redis==6.2.0
referencing==0.37.0
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rq==2.3.3
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.23.0
zstandard==0.25.0
